requirements.txt
Django
django-ckeditor
pillow
djangorestframework
django-filter